+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version 0.2.32
+- Changed how mouseing over status effect titles are displayed within the renderTokenHUD
+
 ## Version 0.2.31
 - Early version of automatic damage calculation implementation, when a actor is selected and a suer right clicks on a roll, the apply damage options will now take the actors Resistances, Weaknesses, and immunities into account when apply damage.
 - Added an additional option denoted as 'true damage' which will ignore all Resistances, Weaknesses, and immunities calculations.